1 - Adjusted Siren's Heart and renamed to Aquatic Heart. 2 - Deleted alt Siren Heart. 3 - Removed discount effect from Cadance Potion. 4 - Ultra Liquidator now takes 20 cursed flames or ichor instead of the Everglade Spray. 5 - Rev+ DoG now gets faster the further he is away from you. 6 - Aquatic Aberration and Parasea hitbox changes. 7 - Leviathan and Siren AI rework. 8 - Siren renamed to Anahita. 9 - Old Duke Tooth Balls get faster the further they are away from you. 10 - Storm Weaver phase 2 charge gets faster the further you are away from it. 11 - Storm Weaver stops charging once it has passed its target a ways. 12 - Yharon second phase issues fixed. 13 - Dragonfolly charge phase speed reduced. 14 - Dragonfolly final phase auras pushed further apart. 15 - Slime God abyss balls fade away instead of blinking out of existence.

+++ Projectiles/Magic/FabRay.cs
@@ -39,7 +39,7 @@ namespace CalamityMod.Projectiles.Magic
 				Main.dust[pink].noGravity = true;
 				Main.dust[pink].noLight = true;
 				Main.dust[pink].position = source;
-				Main.dust[pink].scale = (float)Main.rand.Next(70, 110) * 0.013f;
+				Main.dust[pink].scale = Main.rand.Next(70, 110) * 0.013f;
 				Main.dust[pink].velocity *= 0.1f;
             }
         }
